Move TransformLog to transform/log.js
