+++ lib/orbit-common/cache.js
@@ -5,7 +5,7 @@ import { eq } from 'orbit/lib/eq';
 import CacheIntegrityProcessor from './cache/operation-processors/cache-integrity-processor';
 import SchemaConsistencyProcessor from './cache/operation-processors/schema-consistency-processor';
 import Transform from 'orbit/transform';
-import TransformLog from 'orbit/transform-log';
+import TransformLog from 'orbit/transform/log';
 import Query from 'orbit/query';
 import QueryEvaluator from 'orbit/query/evaluator';
 import QueryOperators from './cache/query-operators';
