+++ lib/orbit/transform/log.js
