+++ lib/orbit/transformable.js
@@ -2,7 +2,7 @@ import Orbit from './main';
 import Evented from './evented';
 import { extend } from './lib/objects';
 import Transform from './transform';
-import TransformLog from './transform-log';
+import TransformLog from './transform/log';
 
 export default {
   /**
